+++ cloud_client/src/main/resources/client_panel.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<VBox xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.geekbrains.cloud_storage.client.ClientController">
+<VBox xmlns="http://javafx.com/javafx/10.0.2-internal" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.geekbrains.cloud_storage.client.controllers.ClientController">
     <HBox>
         <TextField fx:id="pathField" editable="false" HBox.hgrow="ALWAYS" />
         <Button onAction="#btnPathUpAction" text="Вверх" />
